- Correção da lógica de recuperar informações da API e salvar no banco de dados - Remoção do atributo ID_PRODUCT do PhotoContract - Inserção do atributo id_user no objeto Product - Implementação do metodo select no model Photo - Correção do método insert no model Product - Correção do método insert no model User - Implementação do metodo select no model User - Criação da tabela ProductPhoto

+++ app/src/main/java/br/com/enjoeichallenge/repository/managers/SQLiteManager_CRUD.java
@@ -14,7 +14,7 @@ public interface SQLiteManager_CRUD {
     long update(Object obj);
 
     // Select
-    Object select(int id);
+    Object select(long id);
     ArrayList<Object> selectAll();
 
     // Auxiliares
